Moved ClusterTable and PSW into their own files
