+++ src/subCommand/compare/robinsonFoulds.cpp
@@ -3,7 +3,6 @@
 #include <algorithm>
 #include <cstdint>
 #include <cstdlib>
-#include <functional>
 #include <iostream>
 #include <limits>
 #include <stack>
@@ -13,148 +12,16 @@
 #include <utility>
 #include <vector>
 
+#include "util/clusterTable.h"
 #include "util/lap.h"
-
-using PSW = std::vector<std::pair<uint64_t, uint64_t>>;
-
-struct LRNW {
-    uint64_t L, R, N, W;
-};
-
-struct ClusterTable {
-public:
-    ClusterTable(const Graph &g, const PSW &psw) {
-        uint64_t leafCode = 0;
-        internalLabels.reserve(g.leafName.size());
-
-        uint64_t rightLeaf = 0;
-
-        for (size_t i = 0; i < psw.size(); i++) {
-            const std::pair<uint64_t, uint64_t> p = psw[i];
-
-            // If leaf
-            if (p.second == 0) {
-                std::string leafName = g.leafName.at(p.first);
-                internalLabels[leafName] = leafCode;
-
-                rightLeaf = leafCode;
-                leafCode++;
-            } else {
-                uint64_t leftLeafIndex = psw[i - p.second].first;
-                std::string leafName = g.leafName.at(leftLeafIndex);
-                uint64_t leftLeaf = internalLabels[leafName];
-
-                ct[leftLeaf].insert(rightLeaf);
-            }
-        }
-
-    }
-
-    uint64_t encode(const std::string &l) {
-        return internalLabels.at(l);
-    }
-
-    bool isClust(uint64_t L, uint64_t R) {
-        auto itL = ct.find(L);
-
-        if (itL == ct.end()) {
-            return false;
-        }
-
-        return itL->second.find(R) != itL->second.end();
-    }
-
-public:
-    std::unordered_map<uint64_t, std::unordered_set<uint64_t>> ct;
-    std::unordered_map<std::string, uint64_t> internalLabels;
-};
-
-// Post order sequence with weights (PSW)
-static PSW genPSWHelper(
-    const Graph &g,
-    const std::unordered_map<uint64_t, uint8_t> &curEdges
-) {
-    PSW res;
-
-    std::function<uint64_t(uint64_t)> postOrder = [&](uint64_t node) -> uint64_t {
-        uint64_t weight = 0;
-
-        auto leafIt = g.leafName.find(node);
-
-        if (leafIt == g.leafName.end()) {
-            for (const uint64_t &c : g.adjList[node]) {
-                auto it = g.reticulations.find(c);
-
-                // Check if the current node has an edge
-                // to a reticulation at child c
-                if (it != g.reticulations.end()) {
-                    // Child c is a reticulation, so get its parents
-                    const std::vector<uint64_t> &parents = it->second;
-                    uint8_t curEdge = curEdges.at(c);
-
-                    // Continue if there's "no" edge from reticulation's parent
-                    // to the reticulation c
-                    if (parents[curEdge] != node) {
-                        continue;
-                    }
-                }
-
-                weight += postOrder(c);
-            }
-        }
-
-        res.push_back(std::make_pair(node, weight));
-        return weight + 1;
-    };
-
-    postOrder(g.root);
-
-    return res;
-}
-
-static std::vector<PSW> genPSWs(
-    const Graph &g
-) {
-    std::vector<PSW> psw;
-
-    std::unordered_map<uint64_t, uint8_t> curEdges;
-    curEdges.reserve(g.reticulations.size());
-
-    for (const auto &p : g.reticulations) {
-        curEdges[p.first] = 0;
-    }
-
-    while (true) {
-        psw.emplace_back(genPSWHelper(g, curEdges));
-
-        auto it = curEdges.begin();
-        while (it != curEdges.end()) {
-            if (it->second < g.reticulations.at(it->first).size() - 1) {
-                it->second++;
-                break;
-            } else {
-                it->second = 0;
-            }
-
-            it++;
-        }
-
-        if (it == curEdges.end()) {
-            break;
-        }
-    }
-
-    return psw;
-}
+#include "util/psw.h"
 
 // COMCLUST
-static uint64_t rf_dist(
-    const Graph &g1, const PSW &psw1,
+static uint64_t rfDist(
+    const ClusterTable &ct,
     const Graph &g2, const PSW &psw2
 ) {
-    uint64_t res = 0;
-
-    ClusterTable ct(g1, psw1);
+    uint64_t res;
 
     std::stack<LRNW> s;
