+++ src/subCommand/compare/util/psw.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "../../../graph.h"
+
+using PSW = std::vector<std::pair<uint64_t, uint64_t>>;
+
+std::vector<PSW> genPSWs(const Graph &g);
